Feat: 기업회원/개인회원 구분 임의 미들웨어 추가

+++ boards/boards.service/deleteBorad.js
@@ -1,19 +1,27 @@
 const Board = require('../../sequelize/models/board');
 
-module.exports = deleteBoard = async(req, res) =>{
-    const board_id = req.body.board_id;
+module.exports = deleteBoard = async (req, res) => {
+    const deleteObj = req.body;
 
-    try{
+    const company_id = deleteObj.company_id;
+    const board_id = parseInt(deleteObj.board_id);
+
+    //개인 회원 접근 x, 다른 회사 접근 x
+    if (req.id.code === 0 || req.id.company_id != company_id) {
+        return res.status(401).json({ success: false, message: "접근 권한 없음" })
+    }
+
+    try {
         const deleteResult = await Board.destroy({
             where: {
                 board_id: board_id
             }
         });
         console.log("deleteResult", deleteResult);
-        res.status(201).json({success: true, message: "해당공고 삭제완료"})
+        return res.status(201).json({ success: true, message: "해당공고 삭제완료" })
     }
-    catch(e){
+    catch (e) {
         console.log("deleteBoard Err : ", e);
-        res.status(400).json({success: false, message: "잘못된 요청"});
+        return res.status(400).json({ success: false, message: "잘못된 요청" });
     }
 }
